Aula 'Configurando supertest' (C4-M1-SM5)
