+++ src/modules/cars/useCases/createCategory/CreateCategoryController.spec.ts
@@ -1,0 +1,9 @@
+import request from "supertest";
+
+import { app } from "@shared/infra/http/app";
+
+describe("Create Category Controller", () => {
+  it("test", async () => {
+    await request(app).get("/cars/available").expect(201);
+  });
+});
